refactor: move entry previews to their own file
