+++ LabelWorkshop/views/EntryMiniView.swift
@@ -10,30 +10,7 @@ struct EntryMiniView: View {
     var body: some View {
         NavigationLink(destination: EntryView(entry: entry).id(entry.id)){
             VStack(spacing: 0) {
-                if let image = loadImage(for: entry) {
-                    Image(uiImage: image)
-                        .resizable()
-                        .aspectRatio(contentMode: .fill)
-                        .frame(
-                            minWidth: 0,
-                            maxWidth: .infinity,
-                            minHeight: 0,
-                            maxHeight: .infinity
-                        )
-                        .aspectRatio(1 / 1, contentMode: .fit)
-                        .clipShape(Rectangle())
-                        .cornerRadius(8)
-                } else {
-                    Image(systemName: "exclamationmark.triangle.fill")
-                        .symbolRenderingMode(.multicolor)
-                        .font(.system(size: 32))
-                        .frame(
-                            minWidth: 0,
-                            maxWidth: .infinity,
-                            minHeight: 0,
-                            maxHeight: .infinity
-                        )
-                }
+                EntryPreView(entry: entry, square: true)
                 Text(entry.path)
                     .font(.caption)
                     .lineLimit(1)
