+++ LabelWorkshop/views/LibraryView.swift
@@ -1,17 +1,5 @@
 import SwiftUI
 
-func loadImage(for entry: Entry) -> UIImage? {
-    guard let path = entry.fullPath else { return nil }
-    guard let bookmark = entry.library.bookmark else { return nil }
-    guard bookmark.startAccessingSecurityScopedResource() else { return nil }
-    defer { bookmark.stopAccessingSecurityScopedResource() }
-    if let data = try? Data(contentsOf: path) {
-        let uiImage = UIImage(data: data)
-        return uiImage
-    }
-    return nil
-}
-
 struct LibraryView: View {
     let library: Library
